Add test files for service layer

+++ Projects/Project 1/Backend/socialnetwork/src/main/java/com/revature/socialnetwork/service/ReactionService.java
@@ -24,7 +24,7 @@ public interface ReactionService {
      * @param postId the id of the post
      * @return a list of reactions for the specified post
      */
-    List<Reaction> getReactionsByPostId(Integer postId);
+    List<Reaction> getReactionsByPostId(int postId);
 
     /**
      * Retrieves reactions by comment id
@@ -32,7 +32,7 @@ public interface ReactionService {
      * @param commentId the id of the comment
      * @return a list of reactions for the specified comment
      */
-    List<Reaction> getReactionsByCommentId(Integer commentId);
+    List<Reaction> getReactionsByCommentId(int commentId);
 
     /**
      * Retrieves a reaction by its id
@@ -40,14 +40,14 @@ public interface ReactionService {
      * @param id the id of the reaction
      * @return the reaction with the given id
      */
-    Reaction getReactionById(Integer id);
+    Reaction getReactionById(int id);
 
     /**
      * Deletes a reaction by its id
      * 
      * @param id the id of the reaction to be deleted
      */
-    void deleteReaction(Integer id);
+    void deleteReaction(int id);
 
     /**
      * Counts reactions by type for a given post
@@ -56,7 +56,7 @@ public interface ReactionService {
      * @param reactionType the type of reaction to count
      * @return the number of reactions of the specified type for the post
      */
-    Long countReactionsByTypeForPost(Integer postId, ReactionType reactionType);
+    int countReactionsByTypeForPost(int postId, ReactionType reactionType);
 
     /**
      * Counts reactions by type for a given comment
@@ -65,5 +65,5 @@ public interface ReactionService {
      * @param reactionType the type of reaction to count
      * @return the number of reactions of the specified type for the comment
      */
-    Long countReactionsByTypeForComment(Integer commentId, ReactionType reactionType);
+    int countReactionsByTypeForComment(int commentId, ReactionType reactionType);
 }
